Revert previous patch that broke vecds.
git-svn-id: https://svn.code.sf.net/p/vecds/code@111 46551c14-c27e-41b6-87a3-467ec993ed05

+++ include/vecds/main_viewer.h
@@ -82,12 +82,12 @@ namespace vecds
     
   signals:
     
-    void SIG_phiRotationChanged(double);
-    void SIG_thetaRotationChanged(double);
-    void SIG_psiRotationChanged(double);
-    void SIG_xMovementChanged(double);
-    void SIG_yMovementChanged(double);
-    void SIG_zMovementChanged(double);
+    void SIG_phiRotationChanged(int);
+    void SIG_thetaRotationChanged(int);
+    void SIG_psiRotationChanged(int);
+    void SIG_xMovementChanged(int);
+    void SIG_yMovementChanged(int);
+    void SIG_zMovementChanged(int);
     
     void SIG_actPoint(QVector3D res);
     void SIG_actPosition(QVector3D res);
